+++ designer-website/Views/Home/Index.cshtml
@@ -4,5 +4,5 @@
 
 <div class="text-center">
     <h1 class="display-4">Welcome</h1>
-    <p>Sign in @Html.ActionLink("NOW", "SignIn", "Auth")</p>
+    <p>Sign in @Html.ActionLink("NOW", "SignUp", "Account")</p>
 </div>
